// src/common/interfaces/redis.type.ts
export type QueueRedis = 'birthday' |
    'fire_me' |
    'user_view_profile_virtual' |
    'month' |
    'week' |
    'online' |
    'neighbor_city' |
    'follow_up_user_no_login' |
    'like_pin_favorites' |
    'follow_up' |
    'force_follow_up' |
    'follow_up_120' |
    'bring_him_back' |
    'follow_up_5_hrs' |
    'follow_up_1_5_DAY' |
    'follow_up_week' |
    'follow_up_month' |
    'test' 

export enum QueenRedisE {
    BIRTHDAY = 'birthday',
    FIREME = 'fire_me',
    USE_VIEW_PROFILE = 'user_view_profile_virtual',
    MONTH = 'month',
    WEEK = 'week',
    ONLINE = 'online',
    NEIGHBOR_CITY = "neighbor_city",
    FOLLOW_UP_USER_NO_LOGIN = 'follow_up_user_no_login',
    LIKE_PIN_FAVORITES = 'like_pin_favorites',
    FOLLOW_UP = 'follow_up',
    FOLLOW_UP_120 = 'follow_up_120',
    BRING_HIM_BACK = 'bring_him_back',

    FOLLOW_UP_5_HRS = 'follow_up_5_hrs',
    FOLLOW_UP_1_5_DAY = 'follow_up_1_5_DAY',
    FOLLOW_UP_WEEK = 'follow_up_week',
    FOLLOW_UP_MONTH = 'follow_up_month',

    UNLIMITED_COINS_24_HRS = 'unlimited_coins_24_hrs',
    FOLLOW_UP_MESSAGE_UNLIMITED = 'follow_up_message_unlimited',

    FORCE_FOLLOW_UP = 'force_follow_up',

    TEST = 'test'

}